Changed type of fuel related variables from long to double.

+++ xpilot/src/server/object.h
@@ -427,14 +427,14 @@ union _anyobject {
  * Fuel structure, used by player
  */
 typedef struct {
-    long	sum;			/* Sum of fuel in all tanks */
-    long	max;			/* How much fuel can you take? */
+    double	sum;			/* Sum of fuel in all tanks */
+    double	max;			/* How much fuel can you take? */
     int		current;		/* Number of currently used tank */
     int		num_tanks;		/* Number of tanks */
-    long	tank[1 + MAX_TANKS];	/* main fixed tank + extra tanks. */
-    long	l1;			/* Fuel critical level */
-    long	l2;			/* Fuel warning level */
-    long	l3;			/* Fuel notify level */
+    double	tank[1 + MAX_TANKS];	/* main fixed tank + extra tanks. */
+    double	l1;			/* Fuel critical level */
+    double	l2;			/* Fuel warning level */
+    double	l3;			/* Fuel notify level */
 } pl_fuel_t;
 
 struct _visibility {
